refactor: move ipfs code to src
